Check for extension before loading SFX sample
Some charts don't specify an extension for their samples ("2" instead of "2.wav"), this crashes the game since it causes an underflow when trying to extract the extension, it tries to substract 4 from the string length (1). This commit fixes that by checking the string is long enough before attempting to extract an extension from it.